refactor: Tags 탭과 관련된 컴포넌트들 Tags 폴더로 이동
1. 기존: 폴더 없이 파일들이 나열되어 있었음.
2. 폴더 하나로 묶어서 관련 파일을 더 찾기 쉽게 함.
