+++ src/pages/Issue/components/Tags/IssueTags.tsx
